# src/NileTechno.API/openapi/niletechno-openapi.yaml
openapi: 3.0.3
info:
  title: NileTechno API
  description: |
    Frontend API Reference for NileTechno E-Commerce API.
    JSON fields use camelCase. Public product/catalog data only includes active ERP products/groups.
  version: 1.0.0
servers:
  - url: https://my-back-end-jayl.onrender.com
    description: Production
  - url: http://localhost:5080
    description: Local Development
...
security:
  - {}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token. Expires after 60 minutes.
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
...
    Coordinates:
      type: object
      properties:
        lat: { type: number, format: double }
        lng: { type: number, format: double }
    OrderItem:
      type: object
      required: [productId, price, quantity]
      properties:
        productId: { type: string }
        name: { type: string }
        price: { type: number, format: double }
        quantity: { type: integer }
        image: { type: string }
        selectedColor: { type: string, nullable: true }
        selectedSize: { type: string, nullable: true }
    CreateOrderRequest:
      type: object
      description: >
        إجباري: customerName، customerPhone، governorate، عنوان التوصيل
        (أي من address/addressDetails/customerAddress)، items (منتج واحد على الأقل
        بـ productId صحيح وquantity ≥ 1)، total. الخطأ يرجع 400 بصيغة {success:false, errors:[...]}
      required: [customerName, customerPhone, governorate, items, total]
      properties:
        userId: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        customerEmail: { type: string, format: email }
        customerAddress: { type: string }
        address: { type: string }
        addressDetails: { type: string }
        customerNotes: { type: string }
        governorate: { type: string }
        coordinates: { $ref: '#/components/schemas/Coordinates' }
        paymentMethod: { type: string }
        couponCode: { type: string }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/OrderItem' }
        total: { type: number, format: double }
        subtotal: { type: number, format: double }
        shippingCost: { type: number, format: double }
        discountAmount: { type: number, format: double }
    OrderHistoryEntry:
      type: object
      properties:
        action: { type: string }
        actor: { type: string }
        type: { type: string }
        newStatus: { type: string }
        timestamp: { type: string, format: date-time }
    Order:
      type: object
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        userId: { type: string, nullable: true }
        status:
          type: string
          enum: [PENDING, PROCESSING, PACKED, SHIPPED, OUT_FOR_DELIVERY, DELIVERED, CANCELED, REFUNDED]
        cancelReason: { type: string, nullable: true }
        internalNote: { type: string, nullable: true }
        customerName: { type: string }
        customerEmail: { type: string, nullable: true }
        customerPhone: { type: string }
        customerNotes: { type: string, nullable: true }
        governorate: { type: string, nullable: true }
        customerAddress: { type: string, nullable: true }
        addressDetails: { type: string, nullable: true }
        latitude: { type: number, format: double, nullable: true }
        longitude: { type: number, format: double, nullable: true }
        paymentMethod: { type: string, nullable: true }
        couponCode: { type: string, nullable: true }
        discountAmount: { type: number, format: double }
        subtotal: { type: number, format: double }
        shippingCost: { type: number, format: double }
        total: { type: number, format: double }
        date: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        history:
          type: array
          items: { $ref: '#/components/schemas/OrderHistoryEntry' }
    UpdateOrderStatusRequest:
      type: object
      required: [orderId, newStatus]
      properties:
        orderId:
          oneOf:
            - { type: integer }
            - { type: string }
        newStatus:
          type: string
          enum: [PENDING, PROCESSING, PACKED, SHIPPED, OUT_FOR_DELIVERY, DELIVERED, CANCELED, REFUNDED]
        cancelReason: { type: string }
    BulkUpdateOrderStatusRequest:
      type: object
      required: [orderIds, newStatus]
      properties:
        orderIds:
          type: array
          items: { type: integer }
        newStatus:
          type: string
          enum: [PENDING, PROCESSING, PACKED, SHIPPED, OUT_FOR_DELIVERY, DELIVERED, CANCELED, REFUNDED]
        cancelReason: { type: string }
    UpdateOrderNoteRequest:
      type: object
      required: [orderId, note]
      properties:
        orderId:
          oneOf:
            - { type: integer }
            - { type: string }
        note: { type: string }
...
paths:
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List orders
      responses:
        '200':
          description: Orders, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }

    post:
      tags: [Orders]
      summary: Create order
      description: حفظ الطلب في Ec_Orders ثم فتح فاتورة ERP تلقائيًا (wh_TransHeader TransType=3 + wh_TransDetails) فيُخصم المخزون من ERP ويظهر في المتجر خلال دقيقتين. فشل الفاتورة لا يُفشل الطلب (إعادة محاولة تلقائية في الخلفية). الطلب المصحوب بتوكن يُسجَّل باسم حساب التوكن ولا يُستخدم فيه `userId` القادم من الـ body (منعًا لختم سلة عميل آخر في أوردر غريب)، والشراء كزائر بدون توكن يحتفظ بسلوك `userId` من الـ body.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderRequest' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  order: { $ref: '#/components/schemas/Order' }
        '400':
          description: بيانات ناقصة أو منتج غير متاح — {success:false, errors:["اسم العميل مطلوب", "المنتج 'X' غير موجود أو غير متاح حاليًا", ...]}
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  errors:
                    type: array
                    items: { type: string }
  /api/orders/user/{userId}:
    get:
      tags: [Orders]
      summary: List orders of one account
      description: أوردرات عميل محدد — يفتحها صاحب الحساب نفسه أو Admin/MainAdmin فقط، وغير ذلك 403.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Orders (may be empty), newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
        '401': { description: Requires login }
        '403': { description: غير مسموح الاطلاع على أوردرات عميل آخر }
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get one order by id or order number
      description: يقبل `123` أو `ORD-153012-AB12CD` — يفتحها صاحب الأوردر (مطابقته بحسابه أو بإيميله) أو Admin/MainAdmin فقط.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: The order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '401': { description: Requires login }
        '403': { description: غير مسموح الاطلاع على أوردر عميل آخر }
        '404': { description: الطلب غير موجود }
  /api/orders/update-status:
    post:
      tags: [Orders]
      summary: Update order status
      description: عند التحول إلى CANCELED/REFUNDED تُعقَّد فاتورة الـ ERP (Status=0) فيرجع المخزون تلقائيًا، والعودة لحالة نشطة تُلغي التعقيد.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrderStatusRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  order: { $ref: '#/components/schemas/Order' }
                  emailSent: { type: boolean }
        '403': { description: Admin required }
  /api/orders/bulk-update-status:
    post:
      tags: [Orders]
      summary: Bulk update order status
      description: نفس سلوك تعديل الحالة الفردي — الإلغاء يعقّد فاتورة ERP لكل طلب والعودة تنشرها.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkUpdateOrderStatusRequest' }
      responses:
        '200':
          description: Updated orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  updatedCount: { type: integer }
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/Order' }
        '403': { description: Admin required }
  /api/orders/update-note:
    post:
      tags: [Orders]
      summary: Update order internal note
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrderNoteRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  order: { $ref: '#/components/schemas/Order' }
        '403': { description: Admin required }
  /api/orders/{orderId}:
    delete:
      tags: [Orders]
      summary: Delete order
      description: يعقّد فاتورة الـ ERP المرتبطة (إن وُجدت) قبل حذف الطلب من جداول المتجر.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '403': { description: Admin required }
...